+++ extensions/tiktok/src/main/java/app/morphe/extension/tiktok/settings/SettingsStatus.java
@@ -34,6 +34,7 @@ public class SettingsStatus {
     public static boolean commentToolsEnabled = false;
     public static boolean videoOverlaysEnabled = false;
     public static boolean shareSheetEnabled = false;
+    public static boolean ghostModeEnabled = false;
     public static boolean disableTelemetryEnabled = false;
     public static boolean hideFeedFollowButtonEnabled = false;
     public static boolean hideFeedSaveButtonEnabled = false;
@@ -142,6 +143,10 @@ public class SettingsStatus {
 
     public static void enableCommentTools() {
         commentToolsEnabled = true;
+    }
+
+    public static void enableGhostMode() {
+        ghostModeEnabled = true;
     }
 
     public static void enableDisableTelemetry() {